Enable customization of individual node settings
To be able to test Harvester VM scheduling behavior (e.g. schedule
VMs on the nodes that has the available resource), we need to be able
to customize each node's CPU and memory so we can simulation various
resource utilization scenarios.

This patch also enables larger cluster (> 3) nodes and making the node names
static in order to facilitate host-specific testings.