feat: support additional virtualized NVMe disks
Any node for which nvme_size is specified in settings.yml will
automatically have one additional virtual NVMe disk created,
which will appear as /dev/nvme0n1.  This can be useful for
testing additional disks with Longhorn.  The way it works
is a bit obscure because libvirt only seems to support virtio,
scsi and sata disks, so we have to resort to hacking qemu
arguments in as described in this blog post:

  http://blog.frankenmichl.de/2018/02/13/add-nvme-device-to-vm/

This means we have to go create the volume manually first with
`virsh` then pass its path to qemu.  Matters are complicated
further by volumes being owned by root by default, while libvirt
runs VMs as the qemu user.  For normal libvirt volumes, file
ownership changes happen automatically, but for these volumes
we're hacking in by hand, that doesn't happen, so we have to
explicitly specify ownership, and to do _that_ we have to
define the volume via XML, hence the nasty `virsh vol-create-as
... | sed ... | virsh vol-create` invocation.

The other wrinkle is that we need to know the exact path to
the disk image.  Once a volume has been created, you can run
`virsh vol-path --pool default VOLUME_NAME` to get the path,
but we need to know the path when setting libvirt.qemuargs,
whereas we don't want to actually create the volume until the
`vagrant up` trigger for the node in question.  If we create
the volume outside the trigger, it gets created on _every_
evaluation of the Vagrant file, even for unrelated VMs...
So, we call `virsh pool-dumpxml default` and get the pool
path from that, then stick the volume name on the end.

To test, try something like this in settings.yml:

```
harvester_network_config:
  cluster:
    - ip: 192.168.0.30
      # [...]
      nvme_size: 10G
```

Signed-off-by: Tim Serong <tserong@suse.com>