s3: make S3 service store data on extra volume
Make the optional S3 service store data on extra volume with
configurable and sensible size. This prevents it from filling up the
root disk of the DHCP server and allows users to configure it such that
it can take even large and heavy backups.
In addition, there are additional resources allocated to the VM if the
S3 service is enabled to account for the potentially increased resource
usage when backups are taking place.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>